+++ test/reflection_factory_test.dart
@@ -163,6 +163,9 @@ void main() {
               .hasRegisterClassReflection(TestUserWithReflection),
           isTrue);
 
+      expect(userReflection.siblingsClassReflection().map((e) => e.classType),
+          equals([TestUserWithReflection, TestAddressWithReflection]));
+
       expect(
           ReflectionFactory()
               .getRegisterClassReflection(TestUserWithReflection)!